infrastructure/api-gateway/requirements.txt
azure-functions
azure-servicebus
azure-cosmos
pyyaml
